pyproject.toml
[tool.poetry]
name = "ble-rtls-api"
version = "0.1.0"
description = ""
authors = ["Andre Lara <devandrelara@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
SQLAlchemy = "^1.4.40"
psycopg2 = "^2.9.3"
pydantic = "^1.10.0"
paho-mqtt = "^1.6.1"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
pre-commit = "^2.20.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
